requirements.txt
matplotlib
numpy
scipy
tqdm
PyYAML